backend-extraction/pyproject.toml
[project]
name = "backend-extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docling>=2.87.0",
    "fastapi>=0.135.3",
    "python-multipart>=0.0.26",
    "uvicorn>=0.44.0",
    "gspread>=6.1.4",
    "google-auth>=2.38.0",
    "python-dotenv>=1.0.1",
]
